notes_api/requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==18.9b0
Click==7.0
Django==2.2rc1
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
entrypoints==0.3
flake8==3.7.7
mccabe==0.6.1
Pillow==5.4.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
pytz==2018.9
sqlparse==0.3.0
toml==0.10.0
